Add annotations
Add @ExpectWarning annotation with true of false.
Helps to identify faster whether we assume that this example
should be reported or not.
The assumption is based on the question whether this particular
query can result in a SQL injection. A consequence is that
as soon as the second parameter is somehow constant-like we
assume that it is not a SQL injection.